+++ CHANGELOG.md
@@ -2,4 +2,7 @@
 - *Added MIT License*
 - *Updated README*
 - Auto reload config on save
-- 
+- *Added Todo*
+- *Removed extra editor for when the explorer is not visible; updated the code to handle this in only one editor*
+- merged all explorer configs into one tag; also added new config default explorer width
+- *added new dependency `react-terminal`*